Remove define of |WIN32|.
If |WIN32| is going to be defined at all, then it should be defined for
both 32-bit and 64-bit configurations. But, it doesn't need to be
defined at all as long as code is consistently tests |_WIN32|, which is
the proper variable to test. (Whether or not |WIN32| is defined depends
on whether the build system explicitly instructs the compiler to define
it, whereas MSVC and clang-cl automatically define _WIN32.)