Add the |TARGET_OS| build option to facilitate Mac OS X builds.
Carl Mehner reported that the build fails on Mac OS X using its
normal Apple Clang toolchain because that toolchain requires a slightly
different set of perlasm and linker options.

This commit introduces the |TARGET_OS| build option to allow us to make
operating-system-dependent choices during the build. It isn't ideal to
use |TARGET_OS| for this problem because it's really an issue with the
toolchain (a host issue), but this seems to be a reasonable first step
while we figure out a long-term solution.