Add |EC_KEY_new_ex| and |EC_GROUP_new_p256| & similar.
|EC_KEY_new_by_curve_name| and |EC_GROUP_new_by_curve_name| force all
curve implementations to be linked in, even when the application only
wants to use a subset of the available curves. The new functions follow
the pattern used elsewhere where each curve is directly accessed via a
method that doesn't involve a lookup in a table of all curves. When
these new functions are used instead of the old ones, exclusively, only
the implementations for actually-used curves will be linked in, given
appropriate compiler and linker settings.