Require input nonce lengths to equal nominal nonce lengths.
Change the implementation of the C AEAD interface to enforce the
requirement that the nonce length is equal to the result of
|EVP_AEAD_nonce_length| for all AEADs. Previously, the requirement was
not enforced for AES-GCM cipher suites. This in turn enforces that the
input nonces for |open_in_place| and |seal_in_place| are equal to
|ring::aead::Algorithm::nonce_len|.

An issue was filed to remove the underlying code that supports oddball
nonce lengths in the AES-GCM code:
https://github.com/briansmith/ring/issues/22.

In order to move the enforcement of the nonce length to the common code
paths shared by all AEADs, the interface for the AES key wrap AEADs
(currently only available through the C API, not the Rust API) was
changed as described in openssl/include/aead.h.

Some updates were need to keep the C version of the AEAD tests in
crypto/cipher/aead_test.cc working. In particular, the tests were
reorganized so that the tests for |EVP_AEAD_CTX_open| don't depend on
the output of |EVP_AEAD_CTX_seal|, since |EVP_AEAD_CTX_seal| won't
succeed when an invalid nonce is given to it, but we need the test to
keep on going so that we can verify that |EVP_AEAD_CTX_open| also
rejects the invalid nonce. In turn, because we changed the
interpretation of the FAILS attribute in these tests, the old logic
that supported the tests for the removed-in-*ring* stateful AEADs
was removed.