Move |webpki::input| and part of |webpki::der| to *ring*.
The imported code is from libwebpki revision 47629c7. libwebpki will be
refactored to remove its copy of the code in favor of using *ring*'s.

All of |webpki::input| was imported. In the future, it is likely that
some inputs to *ring* APIs will be specified in terms of |ring::input|.

The parts of |webpki::der| that are needed internally within *ring* to
parse public keys values and signatures were also imported. However,
stuff that is not needed internally by *ring* was not imported and will
remain in libwebpki.