Silence older rustc warnings about `usize` in FFI declarations.
Until Rust 1.4, rustc will warn when `usize` or `isize` is used in an
FFI declaration. This patch silences those warnings so that *ring* can
compile without warnings with rustc 1.3 (the current stable version).
This workaround will be reverted when Rust 1.4 is released.

The code doesn't actually trigger this warning yet, but the next commit
would have made it do so.