Add |ring::hmac::verify_with_own_key|.
|ring::hmac::verify_with_own_key| takes a |SigningKey| instead of a
|VerificationKey|. This is more efficient than constructing a
|VerificationKey| with the same key value and then using `verify`. It
also better matches the mental model for the type of applications that
need such functionality; the need to construct two different keys with
the same value was confusing.

Also, the documentation now more clearly explains how to use the HMAC
API for the three different use cases I could think of today.