Remove crypto/ec/example_mul.c.
This code seems to test weird things that I don't understand. I think
it might be some very important regression test for some very important
bug that was fixed, or maybe it is a special test for p256-64.c or
something. Regardless, it was in the way removing it gets it out of the
way. This may not be the best course of action; if not, then it would
be great to learn what this is for.