Fix digest algorithm ID type confusion in |ring::rsa|.
|verify_rsa_pkcs1_signed_digest_asn1| would refuse to consider any
signature valid because it was expecting a NID to identify the
algorithm, but the algorithm ID was actually a |ring::digest::ID|.
The values in each enumeration are completely disjoint.

This would have been caught by the compiler if not for the
`[allow(improper_ctypes)]` workaround for pre-1.4 versions of rustc.
That workaround will be removed in the next commit. This also would
have been caught if the libwebpki tests were run after the commit that
broke this. This would also have been caught if there were RSA signature
verification tests in *ring* (issue #16).