Cherry-pick relevant portions of old commit BoringSSL 231cb82.
Originally *ring* did not merge any portion of BoringSSL 231cb82 as
*ring* does not need to support badly-formed Estonian smart card public
keys. However, instead of ignoring the entire commit, we should take
the parts that verify that |BN_cbs2unsigned| rejects them. Also, it is
a good idea to take the whitespace changes in bn_asn1.c so that it is
easier to compare *ring* and BoringSSL.