Compile with |-fvisibility=hidden|; remove |BORINGSSL_SHARED_LIBRARY|.
The only exported API is the Rust API, and the C code is always linked
statically to the Rust code, so there's never any reason to export any
of the C symbols.

This should allow a dynamic library that has *ring* linked into it to
be used in a program that also links to other OpenSSL forks.