Remove dead code AES code.
BoringSSL uses an approach to platform-specific code that goes to quite
great lengths to avoid conditional preprocessor logic in functions.
However, that approach gets in the way for certain static analysis work
like dead code detection. The use of |abort()| is also problematic as
this code needs to work in contexts where |abort()| is not available.
Accoridngly, get rid of all the callers of |abort()| that are
intentionally dead code.