Remove P-224 & P-521 curves. Fix cross-compilation.
`#[cfg(target_pointer_width)]` in a build script is the host pointer
width, not the end-target pointer width, because during the build of
the build script, the target is the host. Due to this confusion, the
calculation of the constants for P-224 was wrong when the
`sizeof(BN_ULONG)` on the host did not match `sizeof(BN_ULONG)`. To
solve this, limit the curves we support to P-256 and P-384, where
`sizeof(BN_ULONG)` doesn't affect the calculations.