Enable more lints and make them fatal.
Basically, all the lints output by `rustc -W help` were added to
`#![deny(...)]` lists in build.rs and lib.rs. Lints that currently
fail because we do not want to follow the pattern they advise were
added to `#![allow(...)]` lists. Lints that currently fail due to
(minor) problems with the code were also added to the `#![allow(...)]`
lists with `// TODO` comments. Each of those TODOs will be addressed in
the following commits.