Don't overwrite inputs/outputs when AEAD open/seal fails.
This idea was copied from BoringSSL but IMO it isn't so necessary given
that `Result` is `#[must_use]`. It may be more useful for ring-ffi.

Note, in particular, that this was overwriting the entirety of `in_out`
instead of just the output portion. That might have been just plain
wrong.