Don't use the |RSA| structure for RSA signature verification.
Make the RSA signature verification code slightly more efficient by
avoiding the construction of an |RSA| during verification. More
importantly, now the |RSA| structure is exclusively used for signing,
encryption, and decryption, all of which handle secret data, so the way
it is handled can be optimized for those purposes.