Removing casting & some endian-specific logic in p256-64.c.
Remove casting that might violate C's effective types (strict aliasing)
rules.

In the process of doing so, use endian-neutral code instead of code
that only works on little-endian systems. This wasn't the main purpose
of the bug and it isn't yet known if p256-64 is now endian-neutral.

The casts in the call to `batch_mul` will be removed in a separate
commit.