Clarify lifecycle of `BN_BLINDING`.
In `bn_blinding_update` the condition `b->e != NULL` would never be
true (probably), but the test made reasoning about the correctness of
the code confusing. That confusion was amplified by the circuitous and
unusual way in which `BN_BLINDING`s are constructed. Clarify all this
by simplifying the construction of `BN_BLINDING`s, making it more like
the construction of other structures.

Also, make `counter` unsigned as it is no longer ever negative.