Rewrite SHA-1 in Rust.
Fix #60.

I found a number of SHA-1 implementations in Rust but none of them were
in the public domain or under CC0 or ISC, so I rewrote it based on reading
FIPS 180-4.

As discussed in #60 this does not optimize for speed but keeps the source
code short and simple (for example without manual loop unrolling)
as a first approximation of optimizing for machine code size.