Add AEAD tests for all prefix lengths from 0 to 4096 bytes.
The primary purpose of this change is to increase the test coverage
of the prefix-overwriting aspect of AEAD decryption. Unfortunately,
these tests are too slow in debug builds for all these cases to be
tested, so only enable these additional cases on non-debug builds.

The secondary effect of this change is to make the tests more efficient
by factoring out a big chunk of logic from the inner loop of these
tests.