Split the implementation of `ring::signature` into multiple modules.
RSA signing/verification/encryption/decryption all shares a lot of code
and that shared code will now be in `ring::rsa`. Similarly, ECC signing
and key agreement share a lot of stuff, so it is now in `ring::ec`.

This doesn't add any new functionality like RSA encryption; it just
lays the foundation for it.