Remove crypto/ec/ecdsa_test.cc.
New ECDSA verification tests were added to `ring::ec::ecdsa`, and
ECDSA signing isn't exposed as part of the API, so it is OK to
remove the bad testing that is done here.

In particular, note that ecdsa_test cannot find a problem in the ECC
math that would cause both the signer and verifier to compute the same
(wrong) answer, since the signature generation and signature
verification use the same math routines.