Simplify Poly1305 C implementation and make it more type-safe.
The code had effective type rule violations as the `opaque` buffer
was of type `uint8_t[]` but the code was reading/writing it as a
structure type. Fix that by making the code more statically type-safe
in the case where the C implementation is being used.

Note that this doesn't affect the currently-supported platforms because
they all use the assembly language implementations.