Fix references to NSA Suite B Implementer's Guides.
The reference in the ECDSA code was copy/pasted from the ECDH code,
which was referencing (of course) the ECDH guide. Fix it to refer to
the ECDSA guide.

Also, the NSA reorganized their website and now their ECDH guide is
gone and the ECDSA guide is not available at a stable, fixed, URL,
so remove the links.