On Travis CI, allow failures on Rust Nightly.
It is frequently the case that builds using Rust Nightly fail
because the version of libstd we're downloading was compiled with a
different version of rustc than what we're using to build *ring*
during the run. I have no idea why this is so problematic, or why it
seems worse lately, so just wallpaper around it for now.