Fold src/ec/curves.rs into src/ec/ec_build.rs.
Originally it was thought that curves.rs might be reused by *ring*
proper, as well as by the build system. Also, we thought that curves.rs
might be generated by some too. However, now we know that niether will
happen any time soon, so there's no reason for curves.rs to be
separate.