Avoid storing `H` in `GCM128_CONTEXT`.
`H` is only needed during context initialization. However, some
aasembly code requires there to be 16 bytes of data where `H` was, so
put a dummy `H_unused` member there.

Note that in OpenSSL, there are some implementations of GCM that aren't
in *ring*, which do use `H`. In general, we have to be careful about
importing patches to this code from OpenSSL.