Add test for an syntactically-invalid RSA private key component.
The "Invalid key, halved" vector tests that the parsing of the
`SEQUENCE` wrapping the private key components fails because the length
is larger than the length of the remaining input. *This* test helps us
verify that `PositiveInteger`'s `Drop` implementation doesn't go
horrifically wrong. If we ever get memory leak testing, this will help
with that too.