Implement ECDH agreement in Rust.
This unfortunately required a lot of refactoring. The inversion mod q
is now implemented with using a constant-time algorithm, unlike before,
but the point multiplications themselves are still using variable-time
implementations on most platforms, like before.

Note that the inversion algorithm is constant-time, but the primitive
operations that it is comprised of might not be fully constant-time
(yet).