Refactor ECDH private keygen to be constant time & heap-free.
Avoid using the heap during private key generation. Use constant time
comparisons when verifying that the private key is in range. Move the
logic that doesn't need to be constant-time to Rust, leaving only the
constant-time comparisons in C. Adjust the rest of the internal
private key stuff to take into account these changes.