Rename `signature::RSA_PKCS1_*` to put strength ahead of digest again.
Undo the part of 438d04f3d862c210fd11b03937826c0beb9b6f83 that
reorders the components of the `RSA_PKCS1_*` names, but keep the part
that drops the `_VERIFY` suffix. After merging that commit, I came up
with another idea for resolving the inconsistency that avoids adding
the inconsistency with the naming of the `ECDSA_*` things that I'd like
to try first.