Add tests for ECDSA parsing bugs (#243).
This has two ECDSA test vectors with the same key, captured from
TLS handshakes with Cloudflare.  Both are valid signatures, but
one has an `s` which is shorter than usual.

There are also some unit tests for parse_big_endian_value.

These both show the bug.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.

[See https://github.com/briansmith/ring/issues/243 and the associated
pull request. This is the test case commit from that pull request,
unmodified except for moving the ECDSA test vectors to the top of the
file, so that they aren't in the NIST test vector section. The unit
tests will be cleaned up and improved in the next commit. - Brian]