Fix ECDSA signature parsing bug.
The parser did not always correctly parse ECDSA signatures where the
components of the signature had lengths that are not multiples of the
limb size.

In some cases, the parser would parse the input, but with the wrong
value. This does not seem to be exploitable because it we already
assume an attacker will do whatever it wants to the signature.

Other times, it would panic due to an attempt to read out of the bounds
of the input buffer. This seems like it would, at most, cause DoS due
to the program crashing, though it could possibly be used to trigger
latent bugs in an applicatoin's unwind handling if it tries to catch
panics.

Thanks to Joseph Birr-Pixton for reporting the issue and providing the
test cases (in the previous commit).

I added an additional test case for zero-length inputs, though
`der::positive_integer` would have already rejected such inputs in the
first place. I also adjusted the test cases to work for 32-bit targets.

https://github.com/briansmith/ring/issues/243.