Make Rust `BN_BLINDING` a struct, mirror C struct
In anticipation of testing blinding reuse, we want to be able to detect
field-level updates to the C data. Encode the C fields in our Rust type
so we can observe that the counter is updated correctly.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.