Reuse blinding when RSA signing
Note: the existing test only checks that the blinding counter is
incremented modulo `BN_BLINDING_COUNTER`, which does not imply that the
actual `A` and `Ai` values are updated when the counter is reset.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.