Add negative test cases for PBKDF2 and tweak latest PBKDF2 update.
The constant-timedness isn't as strict as I'd like it to be--ideally it
would use the `constant_time_xxx` primitives in C. However, there's not
enough time to create a Rust interface for those functions and to
expand the `ring::constant_time` API to expose functions that operate
on masks, so instead, just approximate what those functions do using
Rust bitwise operations.

Make the `pbkdf2::verify` API stricter by making it reject empty
values, which are nonsensical.

Add negative test vectors.

Simplify the definition of `be_u8_from_u32` and clarify precedence and
masking.