Use the v140, not v140-xp toolchain on Windows.
The v140-xp toolchain is an optional component and there was a report
from somebody that didn't have it installed. Also, we don't have
testing on Windows XP yet, and it's likely we're depending on libstd
features that don't work on Windows XP. In particular, does
`lazy_static` work on XP?

We can always add support for v140-xp later if/when somebody needs it
and we can test it.