Change poly1305 interface to use complete blocks.
The GCM code already works in a similar way. If we need a generic
I-U-F interface for Poly1305 in the future, we'll probably need one for
GCM too, so we'd probably build the I-U-F logic at a higher layer so
tha it could be shared by both. It is likely we won't do that any time
soon.

This also avoids memory copies for aligning the opaque state between
each function call. Now there are no copies.